Misc bug fixes and small improvement
Pulled changes from the BOUT++ repository. Misc bug fixes and
small changes to handle corner cases. Only functional change is to
add a checkbox to force the mesh generator to use a simplified
boundary geometry. This is handy for some ITER discharges where the upper
x-point is very close to the upper divertor plate.